follow SOLID principles
