+++ exercise1/internal/handler/interface.go
@@ -9,7 +9,9 @@ import (
 type ProductService interface {
 	AddProduct(ctx *gofr.Context, product *models.Product) (int, error)
 	GetProduct(ctx *gofr.Context, product *models.Product, id int)(models.Products, error)
+}
 
+type VariantService interface{
 	AddVariant(ctx *gofr.Context, variant *models.Variant) (int, error)
 	GetVariant(ctx *gofr.Context, productID, variantID string)(*models.Variant, error)
 }
