+++ exercise1/internal/store/product/product.go
@@ -1,4 +1,4 @@
-package store
+package product
 
 import (
 	"net/http"
@@ -53,27 +53,4 @@ func (p *ProductRepo) GetProductVariants(ctx *gofr.Context, pid int) []models.Va
 		variants = append(variants, variant)
 	}
 	return variants
-}
-
-func (p *ProductRepo) AddVariant(ctx *gofr.Context, variant *models.Variant) (int, error) {
-	query := "insert into variant(product_id, variantName, variantDetails) values (?,?,?);"
-	res, err := ctx.DB().ExecContext(ctx, query, variant.ProductID, variant.VariantName, variant.VariantDetails)
-	if err != nil {
-		ctx.Logger.Errorf("Error while inserting variant: %v", err)
-		return 0, &errors.Response{
-			StatusCode: http.StatusInternalServerError,
-			Reason:     "oops!! something went wrong...! error encountered in db connection. please try after sometime",
-		}
-	}
-	uId, _ := res.LastInsertId()
-	return int(uId), nil
-}
-
-func (p *ProductRepo) GetVariant(ctx *gofr.Context, productID, variantID string) *models.Variant {
-	var variant models.Variant
-	err := ctx.DB().QueryRowContext(ctx, "select id, product_id, variantName, variantDetails from variant where product_id = ? and id = ?", productID, variantID).Scan(&variant.ID, &variant.ProductID, &variant.VariantName, &variant.VariantDetails)
-	if err != nil {
-		return &models.Variant{}
-	}
-	return &variant
 }
